Move deprecated method DiscoveryService::createClickableFilepath to the new Command class
