+++ packages/framework/src/Console/Commands/RebuildStaticSiteCommand.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Hyde\Console\Commands;
 
 use Exception;
+use Hyde\Console\Concerns\Command as CommandAlias;
 use Hyde\Foundation\Facades\PageCollection;
 use Hyde\Framework\Services\BuildService;
-use Hyde\Framework\Services\DiscoveryService;
 use Hyde\Framework\Services\RebuildService;
 use Hyde\Hyde;
 use LaravelZero\Framework\Commands\Command;
@@ -56,7 +56,7 @@ class RebuildStaticSiteCommand extends Command
 
         $this->info(sprintf(
             'Created %s in %s seconds. (%sms)',
-            DiscoveryService::createClickableFilepath(PageCollection::getPage($this->path)->getOutputPath()),
+            CommandAlias::createClickableFilepath(PageCollection::getPage($this->path)->getOutputPath()),
             number_format(
                 $execution_time,
                 2
