+++ packages/framework/src/Console/Commands/RouteListCommand.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Console\Commands;
 
-use Hyde\Framework\Services\DiscoveryService;
+use Hyde\Console\Concerns\Command as CommandAlias;
 use Hyde\Hyde;
 use LaravelZero\Framework\Commands\Command;
 
@@ -56,7 +56,7 @@ class RouteListCommand extends Command
 
     protected function formatSourcePath(string $path): string
     {
-        return $this->clickablePathLink(DiscoveryService::createClickableFilepath(Hyde::path($path)), $path);
+        return $this->clickablePathLink(CommandAlias::createClickableFilepath(Hyde::path($path)), $path);
     }
 
     protected function formatOutputPath(string $path): string
@@ -65,7 +65,7 @@ class RouteListCommand extends Command
             return "_site/$path";
         }
 
-        return $this->clickablePathLink(DiscoveryService::createClickableFilepath(Hyde::sitePath($path)), "_site/$path");
+        return $this->clickablePathLink(CommandAlias::createClickableFilepath(Hyde::sitePath($path)), "_site/$path");
     }
 
     protected function clickablePathLink(string $link, string $path): string
