+++ packages/framework/src/Console/Concerns/Command.php
@@ -11,5 +11,18 @@ use LaravelZero\Framework\Commands\Command as BaseCommand;
  */
 abstract class Command extends BaseCommand
 {
-    //
+    /**
+     * Create a filepath that can be opened in the browser from a terminal.
+     *
+     * @todo Add support for custom label?
+     * @todo Add option to treat path as already validated so paths that are not created yet can be printed?
+     */
+    public static function createClickableFilepath(string $filepath): string
+    {
+        if (realpath($filepath) === false) {
+            return $filepath;
+        }
+
+        return 'file://'.str_replace('\\', '/', realpath($filepath));
+    }
 }
